Created CSS for the header.
