+++ ScreenDex/src/css/variables.css
@@ -10,7 +10,7 @@
     --font-weight-subheading: 600;
     --font-weight-body: 400;
 
-    --font-size-logo: 100px;
+    --font-size-logo: 50px;
     --font-size-heading: 32px;
     --font-size-subheading: 24px;
     --font-size-body: 16px;
