+++ ScreenDex/src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import Header from '../components/Header';
+import Header from '../components/Header/Header';
 import Footer from '../components/Footer';
 
 function Home() {
