+++ ScreenDex/src/pages/Movies.jsx
@@ -1,4 +1,4 @@
-import Header from '../components/Header';
+import Header from '../components/Header/Header';
 import Footer from '../components/Footer';
 
 import Search from '../components/Search';
